refactor(ui): Separate SPDX and glob schemas into their own files
For reuse, separate the license expression SPDX schema and the glob
pattern schema to common components.
